+++ environment.yml
@@ -10,6 +10,4 @@ dependencies:
   - pytest==6.2.5
   - xarray==0.11.3
   - dash-bootstrap-components==1.1.0
-  - pip
-  - pip:
-    - dash-bio==1.0.2
+  # - dash-bio
